+++ CHANGELOG.md
@@ -12,8 +12,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `rcli` alias for `reduct-cli`, [PR-10](https://github.com/reduct-storage/reduct-cli/pull/10)
 - Options for server URL and API token to alias add
   command, [PR-11](https://github.com/reduct-storage/reduct-cli/pull/11)
-- `rcl bucket ls` to list buckets, [PR-13](https://github.com/reduct-storage/reduct-cli/pull/13)
-- `rcl bucket show` to browse a bucket, [PR-14](https://github.com/reduct-storage/reduct-cli/pull/14)
+- `rcli bucket ls` to list buckets, [PR-13](https://github.com/reduct-storage/reduct-cli/pull/13)
+- `rcli bucket show` to browse a bucket, [PR-14](https://github.com/reduct-storage/reduct-cli/pull/14)
+- `rcli bucket create` to create a bucket, [PR-15](https://github.com/reduct-storage/reduct-cli/pull/15)
 
 ### Changed: